newsletter/requirements.txt
anthropic>=0.25.0
feedparser>=6.0.10
pyyaml>=6.0
resend>=0.7.0
neo4j>=5.19.0
python-dotenv>=1.0.0
requests>=2.31.0
